Validation & exception was passed

+++ SimpleMVC/src/main/java/ru/sd/web/controllers/BookShelfController.java
@@ -4,11 +4,16 @@ import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import ru.sd.app.services.BookService;
 import ru.sd.web.dto.Book;
+import ru.sd.web.dto.BookToFilter;
+import ru.sd.web.dto.BookToRemove;
+
+import javax.validation.Valid;
 
 @Controller
 @RequestMapping(value = "/books")
@@ -16,6 +21,7 @@ public class BookShelfController {
     private final Logger logger = Logger.getLogger(BookShelfController.class);
     private final BookService bookService;
     private static final String BASE_PAGE = "redirect:/books/shelf";
+    private static final String BASIC_VIEW = "book_shelf";
 
     @Autowired
     public BookShelfController(BookService bookService) {
@@ -26,32 +32,69 @@ public class BookShelfController {
     public String books(Model model) {
         logger.info("got book");
         model.addAttribute("book", new Book());
+        model.addAttribute("bookToFilter", new BookToFilter());
         model.addAttribute("filterBook", bookService.getFilter());
+        model.addAttribute("bookToRemove", new BookToRemove());
         model.addAttribute("booklist", bookService.getAllBooks());
-        return "book_shelf";
+
+        return BASIC_VIEW;
     }
 
     @PostMapping("/save")
-    public String saveBook(Book book) {
-        bookService.saveBook(book);
-        logger.info("Current repository size: " + bookService.getAllBooks().size());
-        return BASE_PAGE;
+    public String saveBook(@Valid Book book, BindingResult bindingResult, Model model) {
+        if (bindingResult.hasErrors()) {
+            model.addAttribute("book", book);
+            model.addAttribute("bookToFilter", new BookToFilter());
+            model.addAttribute("filterBook", bookService.getFilter());
+            model.addAttribute("bookToRemove", new BookToRemove());
+            model.addAttribute("booklist", bookService.getAllBooks());
+
+            return BASIC_VIEW;
+        } else {
+            bookService.saveBook(book);
+            logger.info("Current repository size: " + bookService.getAllBooks().size());
+            return BASE_PAGE;
+        }
     }
 
     @PostMapping("/remove")
-    public String removeBook(Book bookToRemove) {
-        if (bookService.removeBook(bookToRemove)) {
-            logger.info("Current repository size: " + bookService.getAllBooks().size());
-        } else logger.info("There is not such book");
-        return BASE_PAGE;
+    public String removeBook(@Valid BookToRemove bookToRemove,
+                             BindingResult bindingResult, Model model) {
+        if (bindingResult.hasErrors()) {
+            logger.warn("bindingResult.hasErrors");
+            model.addAttribute("book", new Book());
+            model.addAttribute("bookToFilter", new BookToFilter());
+            model.addAttribute("filterBook", bookService.getFilter());
+            model.addAttribute("bookToRemove", bookToRemove);
+            model.addAttribute("booklist", bookService.getAllBooks());
+
+            return BASIC_VIEW;
+        } else {
+            if (bookService.removeBook(bookToRemove)) {
+                logger.info("Current repository size: " + bookService.getAllBooks().size());
+            } else logger.info("There is not such book");
+            return BASE_PAGE;
+        }
     }
 
     @PostMapping("/filter")
-    public String setFilterBook(Book bookToFilter) {
-        bookService.setFilter(bookToFilter);
-        logger.info("Filtered list size: " + bookService.getAllBooks().size());
+    public String setFilterBook(@Valid BookToFilter bookToFilter,
+                                BindingResult bindingResult, Model model) {
+        if (bindingResult.hasErrors()) {
+            logger.warn("bindingResult.hasErrors");
+            model.addAttribute("book", new Book());
+            model.addAttribute("bookToFilter", bookToFilter);
+            model.addAttribute("filterBook", bookService.getFilter());
+            model.addAttribute("bookToRemove", new BookToRemove());
+            model.addAttribute("booklist", bookService.getAllBooks());
 
-        return BASE_PAGE;
+            return BASIC_VIEW;
+        } else {
+            bookService.setFilter(bookToFilter);
+            logger.info("Filtered list size: " + bookService.getAllBooks().size());
+
+            return BASE_PAGE;
+        }
     }
 
     @PostMapping("/clear")
